real molwt

+++ ZD_ACID/My_Chem_ml.f90
@@ -179,7 +179,7 @@
 
   type, public :: Chemical 
        character(len=12) :: name
-       integer           :: molwt
+       real              :: molwt
        integer           :: nmhc      ! nmhc (1) or not(0)
        integer           :: carbons   ! Carbon-number
        real              :: nitrogens ! Nitrogen-number
